[refactor] Bookmark UseCase 리팩토링
